Create MiniSnake.c
This is a modified version of MiniPacMan.c. In this program, we create a snake (with length >= 1) instead of a single-element PacMan (length=1). An attempt in creating a snake in this program is to create 2 snakes back-to-back, where the first snake grows by eating the beans and the second "invisible" snake eats the tail of the first snake. This way, the length of the visible snake (=a user input) is kept constant